refactor: move content from index.ts to ThemeModule
